@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace Helium
 {
@@ -30,8 +30,7 @@ namespace Helium
         /// <param name="services">The services in the web host</param>
         public void ConfigureServices(IServiceCollection services)
         {
-            // add MVC 2.2
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddControllers();
 
             // configure Swagger
             services.ConfigureSwaggerGen(options =>
@@ -42,8 +41,16 @@ namespace Helium
             // Register the Swagger generator, defining one or more Swagger documents
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc(Constants.SwaggerName, new Info { Title = Constants.SwaggerTitle, Version = Constants.SwaggerVersion });
+                c.SwaggerDoc(Constants.SwaggerName, new OpenApiInfo { Title = Constants.SwaggerTitle, Version = Constants.SwaggerVersion });
             });
+
+            // add App Insights if key set
+            string appInsightsKey = Configuration.GetValue<string>(Constants.AppInsightsKey);
+
+            if (!string.IsNullOrEmpty(appInsightsKey))
+            {
+                services.AddApplicationInsightsTelemetry(appInsightsKey);
+            }
         }
 
         /// <summary>
@@ -51,7 +58,7 @@ namespace Helium
         /// </summary>
         /// <param name="app"></param>
         /// <param name="env"></param>
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             // differences based on dev or prod
             if (env.IsDevelopment())
@@ -64,6 +71,11 @@ namespace Helium
                 app.UseHsts();
             }
 
+            app.UseRouting();
+
+            // log 4xx and 5xx results to console
+            app.UseLogger(new LoggerOptions { Log2xx = false, Log3xx = false });
+
             // Enable middleware to serve generated Swagger as a JSON endpoint.
             app.UseSwagger();
 
@@ -74,11 +86,16 @@ namespace Helium
                 c.RoutePrefix = string.Empty;
             });
 
-            // handle api and healthz
-            app.UseMvc();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
 
             // use the robots middleware to handle /robots*.txt requests
             app.UseRobots();
+
+            // use the version middleware to handle /version
+            app.UseVersion();
         }
 
         /// <summary>
