@@ -6,36 +6,33 @@
     public class Constants
     {
         public static readonly string SwaggerVersion = "v " + Version.AssemblyVersion;
-        public static readonly string SwaggerName = "helium";
-        public static readonly string SwaggerTitle = "Helium (C#)";
-        public static readonly string SwaggerPath = "/swagger/" + SwaggerName + "/swagger.json";
-        public static readonly string XmlCommentsPath = SwaggerName + ".xml";
+        public const string SwaggerName = "helium";
+        public const string SwaggerTitle = "Helium (C#)";
+        public const string SwaggerPath = "/swagger/" + SwaggerName + "/swagger.json";
+        public const string XmlCommentsPath = SwaggerName + ".xml";
 
-        public static readonly string KeyVaultName = "KeyVaultName";
+        public const string KeyVaultName = "KeyVaultName";
 
-        public static readonly string CosmosUrl = "CosmosUrl";
-        public static readonly string CosmosKey = "CosmosKey";
-        public static readonly string CosmosDatabase = "CosmosDatabase";
-        public static readonly string CosmosCollection = "CosmosCollection";
+        public const string CosmosCollection = "CosmosCollection";
+        public const string CosmosDatabase = "CosmosDatabase";
+        public const string CosmosKey = "CosmosKey";
+        public const string CosmosUrl = "CosmosUrl";
 
-        public static readonly string AppInsightsKey = "AppInsightsKey";
-
-        public static readonly string DALReloadError = "DAL reload failed";
+        public const string AppInsightsKey = "AppInsightsKey";
+        public const string NewKeyLoadedMetric = "newKeyLoaded";
 
         // if port is changed, also update value in the Dockerfiles
-        public static readonly string Port = "4120";
+        public const string Port = "4120";
 
-        public static readonly string HealthzResult = "Movies: 100\r\nActors: 531\r\nGenres: 19";
+        public const int KeyVaultChangeCheckSeconds = 30;
 
-        public static readonly int NotFound = (int)System.Net.HttpStatusCode.NotFound;
-        public static readonly int ServerError = (int)System.Net.HttpStatusCode.InternalServerError;
+        public const string ActorsControllerException = "ActorsControllerException";
+        public const string GenresControllerException = "GenresControllerException";
+        public const string HealthzControllerException = "HealthzControllerException";
+        public const string MoviesControllerException = "MoviesControllerException";
+        public const string FeaturedControllerException = "FeaturedControllerException";
 
-        public static readonly int RotateKeyCheckSeconds = 600;
-        public static int MainLoopSleepMs = 1000;
-
-        public static readonly string ActorsControllerException = "ActorsControllerException";
-        public static readonly string GenresControllerException = "GenresControllerException";
-        public static readonly string HealthzControllerException = "HealthzControllerException";
-        public static readonly string MoviesControllerException = "MoviesControllerException";
+        public const int DefaultPageSize = 100;
+        public const int MaxPageSize = 1000;
     }
 }
