@@ -1,9 +1,10 @@
 ﻿using Helium.DataAccessLayer;
 using Helium.Model;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.Documents;
+using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Logging;
 using System;
+using System.Threading.Tasks;
 
 namespace Helium.Controllers
 {
@@ -13,8 +14,8 @@ namespace Helium.Controllers
     [Route("api/[controller]")]
     public class ActorsController : Controller
     {
-        private readonly ILogger logger;
-        private readonly IDAL dal;
+        private readonly ILogger _logger;
+        private readonly IDAL _dal;
 
         /// <summary>
         ///  Constructor
@@ -24,118 +25,197 @@ namespace Helium.Controllers
         public ActorsController(ILogger<ActorsController> logger, IDAL dal)
         {
             // save to local for use in handlers
-            this.logger = logger;
-            this.dal = dal;
+            _logger = logger;
+            _dal = dal;
         }
 
         /// <summary>
         /// </summary>
         /// <remarks>Returns a JSON array of Actor objects</remarks>
         /// <param name="q">(optional) The term used to search Actor name</param>
-        /// <response code="200">json array of Actor objects or empty array if not found</response>
+        /// <param name="pageNumber">1 based page index</param>
+        /// <param name="pageSize">page size (1000 max)</param>
+        /// <response code="200">JSON array of Actor objects or empty array if not found</response>
         [HttpGet]
         [Produces("application/json")]
-        public IActionResult GetActors([FromQuery] string q)
+        [ProducesResponseType(typeof(Actor[]), 200)]
+        public async Task<IActionResult> GetActorsAsync([FromQuery] string q, [FromQuery] int pageNumber = 1, [FromQuery] int pageSize = Constants.DefaultPageSize)
         {
-            // check the query string
-            if (q == null)
-            {
-                q = string.Empty;
-            }
+            string method = GetMethod(q, pageNumber, pageSize);
 
-            q = q.Trim();
-
-            string method = string.IsNullOrEmpty(q) ? "GetActors" : string.Format("SearchActors:{0}", q);
-
-            logger.LogInformation(method, q);
+            _logger.LogInformation(method);
 
             try
             {
-                return Ok(dal.GetActorsByQuery(q));
+                if (pageSize < 1)
+                {
+                    pageSize = Constants.DefaultPageSize;
+                }
+                else if (pageSize > Constants.MaxPageSize)
+                {
+                    pageSize = Constants.MaxPageSize;
+                }
+
+                pageNumber--;
+
+                if (pageNumber < 0)
+                {
+                    pageNumber = 0;
+                }
+
+                return Ok(await _dal.GetActorsByQueryAsync(q, pageNumber * pageSize, pageSize));
             }
 
-            catch (DocumentClientException dce)
+            catch (CosmosException ce)
             {
-                // log and return 500
-                logger.LogError("DocumentClientException:" + method + ":{0}:{1}:{2}:{3}\r\n{4}", dce.StatusCode, dce.Error, dce.ActivityId, dce.Message, dce);
+                // log and return Cosmos status code
+                _logger.LogError($"CosmosException:{method}:{ce.StatusCode}:{ce.ActivityId}:{ce.Message}\n{ce}");
 
                 return new ObjectResult(Constants.ActorsControllerException)
                 {
-                    StatusCode = Constants.ServerError
+                    StatusCode = (int)ce.StatusCode
+                };
+            }
+
+            catch (System.AggregateException age)
+            {
+                var root = age.GetBaseException();
+
+                if (root == null)
+                {
+                    root = age;
+                }
+
+                // log and return 500
+                _logger.LogError($"AggregateException|{method}|{root.GetType()}|{root.Message}|{root.Source}|{root.TargetSite}");
+
+                return new ObjectResult(Constants.ActorsControllerException)
+                {
+                    StatusCode = (int)System.Net.HttpStatusCode.InternalServerError
                 };
             }
 
             catch (Exception ex)
             {
-                logger.LogError("Exception:" + method + "\r\n{0}", ex);
+                _logger.LogError($"Exception:{method}\n{ex}");
 
                 return new ObjectResult(Constants.ActorsControllerException)
                 {
-                    StatusCode = Constants.ServerError
+                    StatusCode = (int)System.Net.HttpStatusCode.InternalServerError
                 };
             }
         }
 
         /// <summary>
         /// </summary>
-        /// <remarks>Returns a single Actor object by actorId</remarks>
+        /// <remarks>Returns a single JSON Actor by actorId</remarks>
         /// <param name="actorId">The actorId</param>
         /// <response code="404">actorId not found</response>
         [HttpGet("{actorId}")]
         [Produces("application/json")]
         [ProducesResponseType(typeof(Actor), 200)]
         [ProducesResponseType(typeof(void), 404)]
-        public async System.Threading.Tasks.Task<IActionResult> GetActorByIdAsync(string actorId)
+        public async Task<IActionResult> GetActorByIdAsync(string actorId)
         {
-            logger.LogInformation("GetActorByIdAsync {0}", actorId);
+            _logger.LogInformation($"GetActorByIdAsync {actorId}");
 
             try
             {
                 // get a single actor
-                return Ok(await dal.GetActorAsync(actorId));
+                return Ok(await _dal.GetActorAsync(actorId));
             }
 
             // actorId isn't well formed
             catch (ArgumentException)
             {
-                logger.LogInformation("NotFound:GetActorByIdAsync:{0}", actorId);
+                _logger.LogInformation($"NotFound:GetActorByIdAsync:{actorId}");
 
                 // return a 404
                 return NotFound();
             }
 
-            catch (DocumentClientException dce)
+            catch (CosmosException ce)
             {
                 // CosmosDB API will throw an exception on an actorId not found
-                if (dce.StatusCode == System.Net.HttpStatusCode.NotFound)
+                if (ce.StatusCode == System.Net.HttpStatusCode.NotFound)
                 {
-                    logger.LogInformation("NotFound:GetActorByIdAsync:{0}", actorId);
+                    _logger.LogInformation($"NotFound:GetActorByIdAsync:{actorId}");
 
                     // return a 404
                     return NotFound();
                 }
                 else
                 {
-                    // log and return 500
-                    logger.LogError("DocumentClientException:GetActorByIdAsync:{0}:{1}:{2}:{3}\r\n{4}", dce.StatusCode, dce.Error, dce.ActivityId, dce.Message, dce);
+                    // log and return Cosmos status code
+                    _logger.LogError($"CosmosException:GetActorByIdAsync:{ce.StatusCode}:{ce.ActivityId}:{ce.Message}\n{ce}");
 
                     return new ObjectResult(Constants.ActorsControllerException)
                     {
-                        StatusCode = Constants.ServerError
+                        StatusCode = (int)ce.StatusCode
                     };
                 }
+            }
+
+            catch (System.AggregateException age)
+            {
+                var root = age.GetBaseException();
+
+                if (root == null)
+                {
+                    root = age;
+                }
+
+                // log and return 500
+                _logger.LogError($"AggregateException|GetActorByIdAsync|{root.GetType()}|{root.Message}|{root.Source}|{root.TargetSite}");
+
+                return new ObjectResult(Constants.ActorsControllerException)
+                {
+                    StatusCode = (int)System.Net.HttpStatusCode.InternalServerError
+                };
             }
 
             // log and return 500
             catch (Exception e)
             {
-                logger.LogError("Exception:GetActorByIdAsync:{0}\r\n{1}", e.Message, e);
+                _logger.LogError($"Exception:GetActorByIdAsync:{e.Message}\n{e}");
 
                 return new ObjectResult(Constants.ActorsControllerException)
                 {
-                    StatusCode = Constants.ServerError
+                    StatusCode = (int)System.Net.HttpStatusCode.InternalServerError
                 };
             }
         }
+
+        /// <summary>
+        /// Add parameters to the method name if specified in the query string
+        /// </summary>
+        /// <param name="q"></param>
+        /// <param name="pageNumber"></param>
+        /// <param name="pageSize"></param>
+        /// <returns></returns>
+        private string GetMethod(string q, int pageNumber, int pageSize)
+        {
+            string method = "GetActorsAsync";
+
+            if (HttpContext != null && HttpContext.Request != null && HttpContext.Request.Query != null)
+            {
+                // add the query parameters to the method name if exists
+                if (HttpContext.Request.Query.ContainsKey("q"))
+                {
+                    method = string.Format($"{method}:q:{q}");
+                }
+                if (HttpContext.Request.Query.ContainsKey("pageNumber"))
+                {
+                    method = string.Format($"{method}:pageNumber:{pageNumber}");
+                }
+                if (HttpContext.Request.Query.ContainsKey("pageSize"))
+                {
+                    method = string.Format($"{method}:pageSize:{pageSize}");
+                }
+            }
+
+            return method;
+        }
+
     }
 }
