@@ -1,5 +1,6 @@
-using Microsoft.Azure.Documents;
-using System.Linq;
+using Microsoft.Azure.Cosmos;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Helium.DataAccessLayer
 {
@@ -8,18 +9,57 @@ namespace Helium.DataAccessLayer
     /// </summary>
     public partial class DAL
     {
+        const string _genresSelect = "select value m.genre from m where m.type = 'Genre' order by m.genre";
+
         /// <summary>
         /// Read the genres from CosmosDB
         /// </summary>
         /// <returns>List of strings</returns>
-        public IQueryable<string> GetGenres()
+        public async Task<IEnumerable<string>> GetGenresAsync()
         {
             // get all genres as a list of strings
             // the "select value" converts m.genre to a string instead of a document
+            var query = _cosmosDetails.Container.GetItemQueryIterator<string>(new QueryDefinition(_genresSelect), requestOptions: _cosmosDetails.QueryRequestOptions);
 
-            string sql = "select value m.genre from m where m.type = 'Genre' order by m.genre";
-            return client.CreateDocumentQuery<string>(collectionLink, new SqlQuerySpec(sql), feedOptions);
+            List<string> results = new List<string>();
+
+            while (query.HasMoreResults)
+            {
+                foreach (var doc in await query.ReadNextAsync())
+                {
+                    results.Add(doc);
+                }
+            }
+            return results;
         }
 
+        /// <summary>
+        /// Look up the proper Genre by key
+        /// </summary>
+        /// <param name="key"></param>
+        /// <returns>string.Empty or the Genre</returns>
+        public async Task<string> GetGenreAsync(string key)
+        {
+            string sql = string.Format($"select value m.genre from m where m.type = 'Genre' and m.id = '{key.Trim().ToLower()}'");
+
+            var query = _cosmosDetails.Container.GetItemQueryIterator<string>(new QueryDefinition(sql), requestOptions: _cosmosDetails.QueryRequestOptions);
+
+            List<string> results = new List<string>();
+
+            while (query.HasMoreResults)
+            {
+                foreach (var doc in await query.ReadNextAsync())
+                {
+                    results.Add(doc);
+                }
+            }
+
+            if (results != null && results.Count > 0)
+            {
+                return results[0];
+            }
+
+            return string.Empty;
+        }
     }
 }
