@@ -1,5 +1,5 @@
-using System.Linq;
-
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Helium.DataAccessLayer
 {
@@ -8,13 +8,15 @@ namespace Helium.DataAccessLayer
     /// </summary>
     public interface IDAL
     {
-        System.Threading.Tasks.Task<Model.Actor> GetActorAsync(string actorId);
-        IQueryable<Model.Actor> GetActors();
-        IQueryable<Model.Actor> GetActorsByQuery(string q);
-        IQueryable<string> GetGenres();
-        System.Threading.Tasks.Task<Model.Movie> GetMovieAsync(string movieId);
-        IQueryable<Model.Movie> GetMovies();
-        IQueryable<Model.Movie> GetMoviesByQuery(string q);
-        string GetHealthz();
+        Task<Model.Actor> GetActorAsync(string actorId);
+        Task<IEnumerable<Model.Actor>> GetActorsAsync(int offset = 0, int limit = 0);
+        Task<IEnumerable<Model.Actor>> GetActorsByQueryAsync(string q, int offset = 0, int limit = 0);
+        Task<IEnumerable<string>> GetGenresAsync();
+        Task<Model.Movie> GetMovieAsync(string movieId);
+        Task<IEnumerable<Model.Movie>> GetMoviesAsync(int offset = 0, int limit = 0);
+        Task<IEnumerable<Model.Movie>> GetMoviesByQueryAsync(string q, string genre = "", int year = 0, double rating = 0, bool toprated = false, string actorId = "", int offset = 0, int limit = 0);
+        Task<List<string>> GetFeaturedMovieListAsync();
+        Task<Helium.Model.HealthzSuccessDetails> GetHealthzAsync();
+        Task Reconnect(string cosmosUrl, string cosmosKey, string cosmosDatabase, string cosmosCollection, bool force = false);
     }
 }
