@@ -1,4 +1,3 @@
-using Helium;
 using Helium.Controllers;
 using Helium.Model;
 using Microsoft.AspNetCore.Mvc;
@@ -6,24 +5,25 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace UnitTests
 {
     public class ActorsTest
     {
-        private readonly Mock<ILogger<ActorsController>> logger = new Mock<ILogger<ActorsController>>();
-        private readonly ActorsController c;
+        private readonly Mock<ILogger<ActorsController>> _logger = new Mock<ILogger<ActorsController>>();
+        private readonly ActorsController _controller;
 
         public ActorsTest()
         {
-            c = new ActorsController(logger.Object, TestApp.MockDal);
+            _controller = new ActorsController(_logger.Object, TestApp.MockDal);
         }
 
         [Fact]
-        public void GetAllActors()
+        public async Task GetAllActors()
         {
-            OkObjectResult ok = c.GetActors(string.Empty) as OkObjectResult;
+            OkObjectResult ok = await _controller.GetActorsAsync(string.Empty) as OkObjectResult;
 
             Assert.NotNull(ok);
 
@@ -32,13 +32,12 @@ namespace UnitTests
             Assert.NotNull(ie);
 
             Assert.Equal(AssertValues.ActorsCount, ie.ToList<Actor>().Count);
-
         }
 
         [Fact]
-        public void GetActorsBySearch()
+        public async Task GetActorsBySearch()
         {
-            OkObjectResult ok = c.GetActors(AssertValues.ActorSearchString) as OkObjectResult;
+            OkObjectResult ok = await _controller.GetActorsAsync(AssertValues.ActorSearchString) as OkObjectResult;
 
             Assert.NotNull(ok);
 
@@ -47,15 +46,14 @@ namespace UnitTests
             Assert.NotNull(ie);
 
             Assert.Equal(AssertValues.ActorsSearchCount, ie.ToList<Actor>().Count);
-
         }
 
         [Fact]
-        public async void GetActorByIdPass()
+        public async Task GetActorByIdPass()
         {
             // do not use c.GetActorByIdAsync().Result
             // due to thread syncronization issues with build clients, it is not reliable
-            OkObjectResult ok = await c.GetActorByIdAsync(AssertValues.ActorById) as OkObjectResult;
+            OkObjectResult ok = await _controller.GetActorByIdAsync(AssertValues.ActorById) as OkObjectResult;
 
             Assert.NotNull(ok);
 
@@ -67,19 +65,19 @@ namespace UnitTests
         }
 
         [Fact]
-        public async void GetActorByIdFail()
+        public async Task GetActorByIdFail()
         {
             // this will fail GetPartitionKey
-            NotFoundResult nf = await c.GetActorByIdAsync(AssertValues.BadId) as NotFoundResult;
+            NotFoundResult nf = await _controller.GetActorByIdAsync(AssertValues.BadId) as NotFoundResult;
 
             Assert.NotNull(nf);
-            Assert.Equal(Constants.NotFound, nf.StatusCode);
+            Assert.Equal((int)System.Net.HttpStatusCode.NotFound, nf.StatusCode);
 
             // this will fail search
-            nf = await c.GetActorByIdAsync(AssertValues.ActorById + "000") as NotFoundResult;
+            nf = await _controller.GetActorByIdAsync(AssertValues.ActorById + "000") as NotFoundResult;
 
             Assert.NotNull(nf);
-            Assert.Equal(Constants.NotFound, nf.StatusCode);
+            Assert.Equal((int)System.Net.HttpStatusCode.NotFound, nf.StatusCode);
         }
     }
 }
