@@ -4,25 +4,26 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Xunit;
 
 namespace UnitTests
 {
     public class GenresTest
     {
-        private readonly Mock<ILogger<GenresController>> logger = new Mock<ILogger<GenresController>>();
-        private readonly GenresController c;
+        private readonly Mock<ILogger<GenresController>> _logger = new Mock<ILogger<GenresController>>();
+        private readonly GenresController _controller;
 
         public GenresTest()
         {
-            c = new GenresController(logger.Object, TestApp.MockDal);
+            _controller = new GenresController(_logger.Object, TestApp.MockDal);
         }
 
         [Fact]
-        public void GetGenres()
+        public async Task GetGenres()
         {
 
-            OkObjectResult ok = c.GetGenres() as OkObjectResult;
+            OkObjectResult ok = await _controller.GetGenresAsync() as OkObjectResult;
 
             Assert.NotNull(ok);
 
@@ -31,7 +32,6 @@ namespace UnitTests
             Assert.NotNull(ie);
 
             Assert.Equal(AssertValues.GenresCount, ie.ToList().Count);
-
         }
     }
 }
