@@ -1,4 +1,3 @@
-using Helium;
 using Helium.DataAccessLayer;
 using Helium.Model;
 using Newtonsoft.Json;
@@ -25,12 +24,28 @@ namespace UnitTests
 
         public MockDal()
         {
+            string path = "data/";
+
+            if (!File.Exists(path + "actors.json"))
+            {
+                // support running from VS without copying data
+                path = "../../../" + path;
+            }
+
+            if (!File.Exists(path + "actors.json") ||
+                !File.Exists(path + "genres.json") ||
+                !File.Exists(path + "movies.json"))
+            {
+                Console.WriteLine("Unable to find data files");
+                Environment.Exit(-1);
+            }
+
             // load the data from the json files
-            Actors = JsonConvert.DeserializeObject<List<Actor>>(File.ReadAllText(@"data/actors.json"));
+            Actors = JsonConvert.DeserializeObject<List<Actor>>(File.ReadAllText(path + "actors.json"));
 
-            Movies = JsonConvert.DeserializeObject<List<Movie>>(File.ReadAllText(@"data/movies.json"));
+            Movies = JsonConvert.DeserializeObject<List<Movie>>(File.ReadAllText(path + "movies.json"));
 
-            List<GenreDoc> list = JsonConvert.DeserializeObject<List<GenreDoc>>(File.ReadAllText(@"data/genres.json"));
+            List<GenreDoc> list = JsonConvert.DeserializeObject<List<GenreDoc>>(File.ReadAllText(path + "genres.json"));
 
             Genres = new List<string>();
 
@@ -55,12 +70,12 @@ namespace UnitTests
             throw new ArgumentException("NotFound");
         }
 
-        public IQueryable<Actor> GetActors()
+        public Task<IEnumerable<Actor>> GetActorsAsync(int offset = 0, int limit = 0)
         {
-            return Actors.AsQueryable();
+            return Task<IEnumerable<Actor>>.Factory.StartNew(() => { return Actors; });
         }
 
-        public IQueryable<Actor> GetActorsByQuery(string q)
+        public Task<IEnumerable<Actor>> GetActorsByQueryAsync(string q, int offset = 0, int limit = 0)
         {
             List<Actor> res = new List<Actor>();
 
@@ -74,12 +89,12 @@ namespace UnitTests
                 }
             }
 
-            return res.AsQueryable();
+            return Task<IEnumerable<Actor>>.Factory.StartNew(() => { return res; });
         }
 
-        public IQueryable<string> GetGenres()
+        public Task<IEnumerable<string>> GetGenresAsync()
         {
-            return Genres.AsQueryable<string>();
+            return Task<IEnumerable<string>>.Factory.StartNew(() => { return Genres; });
         }
 
         public Task<Movie> GetMovieAsync(string movieId)
@@ -97,34 +112,120 @@ namespace UnitTests
             throw new ArgumentException("NotFound");
         }
 
-        public IQueryable<Movie> GetMovies()
+        public Task<IEnumerable<Movie>> GetMoviesAsync(int offset = 0, int limit = 0)
         {
-            return Movies.AsQueryable();
+            return Task<IEnumerable<Movie>>.Factory.StartNew(() => { return Movies; });
         }
 
-        public IQueryable<Movie> GetMoviesByQuery(string q)
+        public Task<IEnumerable<Movie>> GetMoviesByQueryAsync(string q, string genre, int year = 0, double rating = 0.0, bool topRated = false, string actorId = "", int offset = 0, int limit = 0)
         {
             List<Movie> res = new List<Movie>();
 
-            q = q.ToLower();
-
-            foreach (Movie m in Movies)
+            if (!string.IsNullOrEmpty(q))
             {
-                if (m.TextSearch.Contains(q))
+                q = q.ToLower();
+
+                foreach (Movie m in Movies)
                 {
-                    res.Add(m);
+                    if (m.TextSearch.Contains(q))
+                    {
+                        res.Add(m);
+                    }
                 }
             }
 
-            return res.AsQueryable();
+            else if (!string.IsNullOrEmpty(genre))
+            {
+                foreach (Movie m in Movies)
+                {
+                    if (m.Genres.Contains(genre, StringComparer.OrdinalIgnoreCase))
+                    {
+                        res.Add(m);
+                    }
+                }
+            }
+
+            else if (year > 0)
+            {
+                foreach (Movie m in Movies)
+                {
+                    if (m.Year == year)
+                    {
+                        res.Add(m);
+                    }
+                }
+            }
+
+            else if (rating > 0)
+            {
+                foreach (Movie m in Movies)
+                {
+                    if (m.Rating >= rating)
+                    {
+                        res.Add(m);
+                    }
+                }
+            }
+
+            else if (topRated)
+            {
+                foreach (Movie m in Movies)
+                {
+                    if (m.Rating >= 8.8)
+                    {
+                        res.Add(m);
+                    }
+                }
+            }
+
+            else if (!string.IsNullOrEmpty(actorId))
+            {
+                actorId = actorId.Trim().ToLower();
+
+                foreach (Movie m in Movies)
+                {
+                    foreach (var a in m.Roles)
+                    {
+                        if (a.ActorId == actorId)
+                        {
+                            res.Add(m);
+                        }
+                    }
+                }
+            }
+
+            else
+            {
+                return Task<IEnumerable<Movie>>.Factory.StartNew(() => { return Movies; });
+            }
+
+            return Task<IEnumerable<Movie>>.Factory.StartNew(() => { return res; });
         }
 
-        public string GetHealthz()
+        public Task<List<string>> GetFeaturedMovieListAsync()
         {
-            return Constants.HealthzResult;
+            List<string> res = new List<string> { "tt0133093", "tt0120737", "tt0167260", "tt0167261", "tt0372784", "tt0172495", "tt0317705" };
+
+            return Task<List<string>>.Factory.StartNew(() => { return res; });
+        }
+
+        public Task<HealthzSuccessDetails> GetHealthzAsync()
+        {
+            HealthzSuccessDetails d = new HealthzSuccessDetails();
+
+            d.Actors = 531;
+            d.Movies = 100;
+            d.Genres = 19;
+
+            return Task<HealthzSuccessDetails>.Factory.StartNew(() => { return d; });
+        }
+
+        public Task Reconnect(string cosmosUrl, string cosmosKey, string cosmosDatabase, string cosmosCollection, bool force = false)
+        {
+            // do nothing
+            return null;
         }
     }
-
 
     public class GenreDoc
     {
